refactoring: extracted some code from StoryLive
